Add DefendantAccountHistoryService and related history sources for account history retrieval
