+++ src/main/java/uk/gov/hmcts/opal/service/opal/history/defendant/sources/EnforcementHistorySource.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.opal.service.opal.history.source;
+package uk.gov.hmcts.opal.service.opal.history.defendant.sources;
 
 import java.time.LocalDate;
 import java.util.List;
@@ -9,20 +9,20 @@ import org.springframework.transaction.annotation.Transactional;
 import uk.gov.hmcts.opal.dto.history.DefendantAccountHistoryFilter;
 import uk.gov.hmcts.opal.dto.history.DefendantAccountHistoryItem;
 import uk.gov.hmcts.opal.dto.history.HistoryItemType;
-import uk.gov.hmcts.opal.entity.paymentterms.PaymentTermsEntity;
-import uk.gov.hmcts.opal.mapper.history.PaymentTermsEntityHistoryMapper;
-import uk.gov.hmcts.opal.repository.PaymentTermsRepository;
+import uk.gov.hmcts.opal.entity.enforcement.EnforcementEntity;
+import uk.gov.hmcts.opal.mapper.history.EnforcementEntityHistoryMapper;
+import uk.gov.hmcts.opal.repository.EnforcementRepository;
 import uk.gov.hmcts.opal.service.opal.history.core.AccountHistoryContext;
 import uk.gov.hmcts.opal.service.opal.history.core.AccountHistorySource;
 import uk.gov.hmcts.opal.service.opal.history.core.AccountHistoryType;
 
 @Service
 @RequiredArgsConstructor
-public class PaymentTermsHistorySourceService extends HistorySourceSpecificationSupport
+public class EnforcementHistorySource extends HistorySourceSpecificationSupport
     implements AccountHistorySource {
 
-    private final PaymentTermsRepository paymentTermsRepository;
-    private final PaymentTermsEntityHistoryMapper paymentTermsEntityHistoryMapper;
+    private final EnforcementRepository enforcementRepository;
+    private final EnforcementEntityHistoryMapper enforcementEntityHistoryMapper;
 
     @Transactional(readOnly = true)
     @Override
@@ -32,33 +32,32 @@ public class PaymentTermsHistorySourceService extends HistorySourceSpecification
 
     @Override
     public HistoryItemType getItemType() {
-        return HistoryItemType.PAYMENT_TERMS;
+        return HistoryItemType.ENFORCEMENT;
     }
 
     @Override
     public List<DefendantAccountHistoryItem> fetch(AccountHistoryContext context,
                                                    DefendantAccountHistoryFilter filter) {
         Long defendantAccountId = context.getAccountId();
-        return paymentTermsRepository.findAll(allOf(
-                paymentTermsForDefendantAccount(defendantAccountId),
-                paymentTermsDateFrom(filter.getDateFrom()),
-                paymentTermsDateTo(filter.getDateTo())
+        return enforcementRepository.findAll(allOf(
+                enforcementForDefendantAccount(defendantAccountId),
+                enforcementDateFrom(filter.getDateFrom()),
+                enforcementDateTo(filter.getDateTo())
             )).stream()
-            .map(paymentTermsEntityHistoryMapper::toHistoryItem)
+            .map(enforcementEntityHistoryMapper::toHistoryItem)
             .toList();
     }
 
-    private Specification<PaymentTermsEntity> paymentTermsForDefendantAccount(Long defendantAccountId) {
-        return (root, query, builder) -> builder.equal(
-            root.get("defendantAccount").get("defendantAccountId"), defendantAccountId);
+    private Specification<EnforcementEntity> enforcementForDefendantAccount(Long defendantAccountId) {
+        return (root, query, builder) -> builder.equal(root.get("defendantAccountId"), defendantAccountId);
     }
 
-    private Specification<PaymentTermsEntity> paymentTermsDateFrom(LocalDate dateFrom) {
+    private Specification<EnforcementEntity> enforcementDateFrom(LocalDate dateFrom) {
         return dateFrom == null ? null
             : (root, query, builder) -> builder.greaterThanOrEqualTo(root.get("postedDate"), atStartOfDay(dateFrom));
     }
 
-    private Specification<PaymentTermsEntity> paymentTermsDateTo(LocalDate dateTo) {
+    private Specification<EnforcementEntity> enforcementDateTo(LocalDate dateTo) {
         return dateTo == null ? null
             : (root, query, builder) -> builder.lessThan(root.get("postedDate"), dayAfterStart(dateTo));
     }
