+++ src/main/java/uk/gov/hmcts/opal/service/opal/history/defendant/sources/PaymentTermsHistorySource.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.opal.service.opal.history.source;
+package uk.gov.hmcts.opal.service.opal.history.defendant.sources;
 
 import java.time.LocalDate;
 import java.util.List;
@@ -9,20 +9,20 @@ import org.springframework.transaction.annotation.Transactional;
 import uk.gov.hmcts.opal.dto.history.DefendantAccountHistoryFilter;
 import uk.gov.hmcts.opal.dto.history.DefendantAccountHistoryItem;
 import uk.gov.hmcts.opal.dto.history.HistoryItemType;
-import uk.gov.hmcts.opal.entity.enforcement.EnforcementEntity;
-import uk.gov.hmcts.opal.mapper.history.EnforcementEntityHistoryMapper;
-import uk.gov.hmcts.opal.repository.EnforcementRepository;
+import uk.gov.hmcts.opal.entity.paymentterms.PaymentTermsEntity;
+import uk.gov.hmcts.opal.mapper.history.PaymentTermsEntityHistoryMapper;
+import uk.gov.hmcts.opal.repository.PaymentTermsRepository;
 import uk.gov.hmcts.opal.service.opal.history.core.AccountHistoryContext;
 import uk.gov.hmcts.opal.service.opal.history.core.AccountHistorySource;
 import uk.gov.hmcts.opal.service.opal.history.core.AccountHistoryType;
 
 @Service
 @RequiredArgsConstructor
-public class EnforcementHistorySourceService extends HistorySourceSpecificationSupport
+public class PaymentTermsHistorySource extends HistorySourceSpecificationSupport
     implements AccountHistorySource {
 
-    private final EnforcementRepository enforcementRepository;
-    private final EnforcementEntityHistoryMapper enforcementEntityHistoryMapper;
+    private final PaymentTermsRepository paymentTermsRepository;
+    private final PaymentTermsEntityHistoryMapper paymentTermsEntityHistoryMapper;
 
     @Transactional(readOnly = true)
     @Override
@@ -32,32 +32,33 @@ public class EnforcementHistorySourceService extends HistorySourceSpecificationS
 
     @Override
     public HistoryItemType getItemType() {
-        return HistoryItemType.ENFORCEMENT;
+        return HistoryItemType.PAYMENT_TERMS;
     }
 
     @Override
     public List<DefendantAccountHistoryItem> fetch(AccountHistoryContext context,
                                                    DefendantAccountHistoryFilter filter) {
         Long defendantAccountId = context.getAccountId();
-        return enforcementRepository.findAll(allOf(
-                enforcementForDefendantAccount(defendantAccountId),
-                enforcementDateFrom(filter.getDateFrom()),
-                enforcementDateTo(filter.getDateTo())
+        return paymentTermsRepository.findAll(allOf(
+                paymentTermsForDefendantAccount(defendantAccountId),
+                paymentTermsDateFrom(filter.getDateFrom()),
+                paymentTermsDateTo(filter.getDateTo())
             )).stream()
-            .map(enforcementEntityHistoryMapper::toHistoryItem)
+            .map(paymentTermsEntityHistoryMapper::toHistoryItem)
             .toList();
     }
 
-    private Specification<EnforcementEntity> enforcementForDefendantAccount(Long defendantAccountId) {
-        return (root, query, builder) -> builder.equal(root.get("defendantAccountId"), defendantAccountId);
+    private Specification<PaymentTermsEntity> paymentTermsForDefendantAccount(Long defendantAccountId) {
+        return (root, query, builder) -> builder.equal(
+            root.get("defendantAccount").get("defendantAccountId"), defendantAccountId);
     }
 
-    private Specification<EnforcementEntity> enforcementDateFrom(LocalDate dateFrom) {
+    private Specification<PaymentTermsEntity> paymentTermsDateFrom(LocalDate dateFrom) {
         return dateFrom == null ? null
             : (root, query, builder) -> builder.greaterThanOrEqualTo(root.get("postedDate"), atStartOfDay(dateFrom));
     }
 
-    private Specification<EnforcementEntity> enforcementDateTo(LocalDate dateTo) {
+    private Specification<PaymentTermsEntity> paymentTermsDateTo(LocalDate dateTo) {
         return dateTo == null ? null
             : (root, query, builder) -> builder.lessThan(root.get("postedDate"), dayAfterStart(dateTo));
     }
